Addresses comments
