+++ src/components/Card/Card.stories.mdx
@@ -1,0 +1,19 @@
+import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
+import * as stories from './Card.stories';
+
+<Meta title='Components/Card/Readme' />
+
+# Card
+
+The `Card` component its a wrapper that will render any content that receives as the following RecentActivity component.
+
+## Props
+
+https://github.com/simpleviewinc/sv-mosaic/blob/develop/src/components/Card/CardTypes.ts
+
+## Recent Activity
+
+### Single Radio Button
+<Preview>
+  <stories.Playground />
+</Preview>
