+++ src/components/Card/Card.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, cleanup, fireEvent } from "@testing-library/react";
 import * as React from "react";
 
 // Components
-import SummaryPageCard from "./SummaryPageCard";
+import SummaryPageCard from "./Card";
 import AddIcon from "@mui/icons-material/Add";
 import ContactsIcon from "@mui/icons-material/Contacts";
