+++ src/components/Card/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./Card";
+export * from "./Card";
+export * from "./CardTypes";
