+++ src/components/RecentActivity/RecentActivity.stories.mdx
@@ -1,0 +1,18 @@
+import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
+import * as stories from './RecentActivity.stories';
+
+<Meta title='Components/RecentActivity/Readme' />
+
+# RecentActivity
+
+The `RecentActivity` component will show the latest actions and interactions with a specific record.
+
+## Props
+
+https://github.com/simpleviewinc/sv-mosaic/blob/develop/src/components/RecentActivity/RecentActivityTypes.ts
+
+## Recent Activity
+
+<Preview>
+  <stories.Example />
+</Preview>
