--- src/components/SummaryPageCard/SummaryPageCard.stories.mdx
@@ -1,24 +1,0 @@
-import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
-import * as stories from './SummaryPageCard.stories';
-
-<Meta title='Components/SummaryPageCard/Readme' />
-
-# Summary Page Card
-
-The `SummaryPageCard` component its a wrapper that will render any content that receives as the following RecentActivity component.
-
-## Recent Activity
-
-The `RecentActivity` component will help to see the latests actions and interactions with a specific record.
-
-## Props
-
-https://github.com/simpleviewinc/sv-mosaic/blob/develop/src/components/SummaryPageCard/SummaryPageCard.tsx
-https://github.com/simpleviewinc/sv-mosaic/blob/develop/src/components/RecentActivity/RecentActivity.tsx
-
-## Recent Activity
-
-### Single Radio Button
-<Preview>
-  <stories.KitchenSink />
-</Preview>
