--- src/components/SummaryPageCard/index.ts
@@ -1,3 +1,0 @@
-export { default } from "./SummaryPageCard";
-export * from "./SummaryPageCard";
-export * from "./SummaryPageCardTypes";
