+++ src/index.ts
@@ -31,6 +31,9 @@ export * from "./components/Button";
 export { default as ButtonRow } from "./components/ButtonRow";
 export * from "./components/ButtonRow";
 
+export { default as Card } from "./components/Card";
+export * from "./components/Card";
+
 export { default as Checkbox } from "./components/Checkbox";
 export * from "./components/Checkbox";
 
@@ -66,6 +69,9 @@ export * from "./components/ToggleSwitch";
 
 export { default as RadioButton } from "./components/RadioButton";
 export * from "./components/RadioButton";
+
+export { default as RecentActivity } from "./components/RecentActivity";
+export * from "./components/RecentActivity";
 
 export { default as Snackbar } from "./components/Snackbar";
 export * from "./components/Snackbar";
